% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{raplogManager}
\alias{raplogManager}
\title{Management of shiny application and report logging at Rapporteket}
\usage{
raplogManager()
}
\value{
Silently exits after successful file operations
}
\description{
A tool to aid archiving and cleanup of logs containing data on use of shiny
application and reports at Rapporteket. This function is ment to be called
from a scheduler (\emph{e.g.} chron) directly or as part of a process chain
initiated by a scheduler
}
\examples{
\donttest{
# Depend on the R_RAP_CONFIG_PATH being defined and that relevant
# configuration is present
raplogManager()
}
}
\seealso{
To become orchestrator
\code{\link[rapbase:runAutoReport]{runAutoReport}} in package \emph{rapbase}
}
