% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{repLogger}
\alias{repLogger}
\title{Logging at report level}
\usage{
repLogger(session, msg = "No message provided",
  .topcall = sys.call(-1), .topenv = parent.frame())
}
\arguments{
\item{session}{Shiny session object to be used for getting user data.
For testing and development purposes \code{session} can be replaced by
\code{list()} in which case various config options might be used to provide
something sensible}

\item{msg}{String providing a user defined message to be added to the log
record. Default value is 'No message provided'}

\item{.topcall}{Parent call (if any) calling this function. Used to provide
the function call with arguments. Default value is \code{sys.call(-1)}}

\item{.topenv}{Name of the parent environment calling this function. Used to
provide package name (\emph{i.e.} register) this function was called from.
Default value is \code{parent.frame()}}
}
\value{
Returns nothing but calls a logging appender
}
\description{
To be used for logging at report level, \emph{i.e.} each time a report is
run. Calls to this function can be made from reactive enviroments within the
shiny server function or from within the (report) functions used by the same
reactive environments
}
\examples{
\donttest{
# Depend on the environment variable R_RAP_CONFIG_PATH being set
repLogger(list())
}
}
