% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{appLogger}
\alias{appLogger}
\title{Log events at application level}
\usage{
appLogger(session, msg = "No message provided")
}
\arguments{
\item{session}{Shiny session object to be used for getting user data.
For testing and development purposes \code{session} can be replaced by
\code{list()} in which case various config options might be used to provide
something sensible}

\item{msg}{String providing a user defined message to be added to the log
record. Default value is 'No message provided'}
}
\value{
Returns nothing but calls a logging appender
}
\description{
To be used for logging at application level \emph{i.e.} when a shiny session
is started.
}
\details{
The below fields will be appended to the application log, in the
following order:
\enumerate{
\item \code{time}: date-time as event is logged as
\code{format(time, "\%Y-\%m-\%d \%H:\%M:\%S")}
\item \code{user}: username as found in the shiny session object
\item \code{name}: full name of user as found in the shiny session object
\item \code{group}: users group membership as provided by the shiny
session object. Normally, this will correspond to the registry the user
belongs to
\item \code{role}: users role as provided by the shiny session object. Its
value will depend on whatever is delivered by the autorization provider,
but for OpenQReg registires 'LU' (local user) and 'SC' (system
coordinator) are typical values
\item \code{resh_id}: the organization id of the current user as provided
by the shiny session object
\item message: an optional message defined as argument to the function
}
}
\examples{
\donttest{
# Depend on the environment variable R_RAP_CONFIG_PATH being set
appLogger(list())
}

}
