% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\alias{createArchive}
\alias{archiveLog}
\alias{cleanArchive}
\title{Archive functions}
\usage{
createArchive(archivePath)

archiveLog(archivePath, logPath, logs = c("appLog", "repLog"))

cleanArchive(archivePath, eolDays)
}
\arguments{
\item{archivePath}{}
}
\description{
Archive functions
}
